test(kit): move Vue message tests into test directory
